Rename share-on to share-button
